random-data/requirements.txt
opensearch-py==2.0.0
mysql-connector-python==8.0.30
python-dateutil==2.8.2
numpy==1.23.1
faker==12.3.3
faker-food==0.1.0
pandas==1.4.3
sqlalchemy==1.4.39